--- src/main/java/racingcar/domain/car/CarDto.java
@@ -1,8 +1,0 @@
-package racingcar.domain.car;
-
-public record CarDto(
-        Name name,
-        Position position
-) {
-
-}
