+++ src/main/java/racingcar/domain/car/dto/CarDto.java
@@ -1,0 +1,11 @@
+package racingcar.domain.car.dto;
+
+import racingcar.domain.car.wrapper.Name;
+import racingcar.domain.car.wrapper.Position;
+
+public record CarDto(
+        Name name,
+        Position position
+) {
+
+}
