+++ src/main/java/racingcar/domain/car/wrapper/Name.java
@@ -1,4 +1,4 @@
-package racingcar.domain.car;
+package racingcar.domain.car.wrapper;
 
 import racingcar.constants.ErrorConsts;
